requirements.txt
apache-airflow==2.9.3
apache-airflow-providers-postgres
apache-airflow-providers-apache-spark
pandas
openpyxl
pyspark
py4j
typing_extensions>=4.0.0
psycopg2-binary
sqlalchemy
scikit-learn
